Add explicit definition of SSIZE_MAX
The specific implementation has been based on [libsodium], which
assumes that SIZE_MAX has been defined. [openssl] seems to define
the SSIZE_MAX constant by hand, according to the word size that
we are currently compiling on. I do not want this, because I
think generic solutions are better. However, I am afraid that
this could ever fail because of an undefined MAX_SIZE. If this
happens we can still adopt [rust]'s approach, which is using bit
twiddling to get a '0b0111...111' value.

[libsodium]: https://github.com/jedisct1/libsodium/blob/1.0.15/src/libsodium/randombytes/sysrandom/randombytes_sysrandom.c#L62-L64
[openssl]: https://mta.openssl.org/pipermail/openssl-dev/2016-March/006412.html
[rust]: https://github.com/rust-lang/rust/blob/1.21.0/src/libcore/num/mod.rs#L104-L128