Suggest edits to #8
I think it is reasonable not to used a separate call to _malloc(),
because we already have a buffer. This patch should write to the
buffer directly, instead of using a temporary buffer.

An important note however, after measuring wall clock time when
filling large buffers with random bytes, I see no difference at
all. Perhaps the indermediate buffer is optimized out (or OpenSSL
is slow)?